pyproject.toml
[project]
name = "challenge-dangerous-worm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cffi==1.15.0",
    "pycryptodome==3.14.1",
    "pyinstaller>=6.16.0",
]
